+++ src/stake/MyERC20.sol
@@ -6,11 +6,9 @@ import "@openzeppelin/ERC20/IERC20.sol";
 import "@openzeppelin/ERC20/extensions/ERC20Permit.sol";
 
 contract MyToken is ERC20 {
-    constructor() ERC20("MyToken", "MTK") {
-     
-    }
+    constructor() ERC20("MyToken", "MTK") {}
 
-    function mint(address add, uint amount) public {
-        _mint(add , amount);
+    function mint(address add, uint256 amount) public {
+        _mint(add, amount);
     }
 }
