+++ test/wallet/wallet3.t.sol
@@ -14,8 +14,8 @@ contract CollectorsTest is Test {
     address public collector2 = address(2);
     address public collector3 = address(3);
     address public randomAddress = address(4);
-    uint val = 10;
-    uint initialValue = 100;
+    uint256 val = 10;
+    uint256 initialValue = 100;
     // OwnerUpOnly public upOnly;
 
     /// @dev initialize everything to the tests
@@ -26,7 +26,9 @@ contract CollectorsTest is Test {
         // address my_add = address(c);
         // upOnly = new OwnerUpOnly();
     }
+
     receive() external payable {}
+
     function testConstructor() public {
         console.log(address(this));
         console.log(address(c));
@@ -34,6 +36,7 @@ contract CollectorsTest is Test {
         // console.log();
         assertEq(c.owner(), address(this));
     }
+
     function testDeposit() public {
         assertEq(0, address(c).balance);
         vm.startPrank(collector1);
@@ -42,47 +45,55 @@ contract CollectorsTest is Test {
         assertEq(val, address(c).balance);
         vm.stopPrank();
     }
+
     function testFailDeposit() public {
         vm.startPrank(randomAddress);
         vm.expectRevert("don't have enough amount");
         c.deposit{value: val}();
     }
+
     function testWithdraw() public {
         //what is the difference between msg.sender and address(this)?
         assertEq(0, address(c).balance);
-        vm.deal(payable(address(c)),initialValue);
+        vm.deal(payable(address(c)), initialValue);
         uint256 firstBalance = address(this).balance;
         console.log(firstBalance);
         c.withdraw(val);
-        assertEq(initialValue-val, address(c).balance);
-        assertEq(firstBalance+val, address(this).balance);
+        assertEq(initialValue - val, address(c).balance);
+        assertEq(firstBalance + val, address(this).balance);
         vm.stopPrank();
     }
+
     function testWithdrawNotAllowed() public {
         vm.startPrank(randomAddress);
         vm.expectRevert("not allowed");
         c.withdraw(val);
     }
+
     function testWithdrawNoMoney() public {
         c.addCollector(collector1);
         vm.startPrank(collector1);
         vm.expectRevert("no money");
         c.withdraw(val);
     }
+
     function testfailAddCollectorNotOwner() public {
         vm.prank(address(123));
         vm.expectRevert("Not owner");
         c.addCollector(collector1);
     }
+
     function testRemoveCollectorNotOwner() public {
         vm.prank(address(123));
         vm.expectRevert("Not owner");
         c.addCollector(collector1);
     }
+
     function testAddCollector() public {
         c.addCollector(collector1);
-        assertEq( c.collector1(), collector1, "add new collector");
+        assertEq(c.collector1(), collector1, "add new collector");
     }
+
     function testAddCollectorNoSpace() public {
         c.addCollector(collector1);
         c.addCollector(collector2);
@@ -91,23 +102,27 @@ contract CollectorsTest is Test {
         vm.expectRevert("couldn't add collector");
         c.addCollector(collector1);
     }
+
     function testAddExistCollector() public {
         c.addCollector(collector1);
         vm.expectRevert("this collector is allowed yet");
         c.addCollector(collector1);
     }
+
     function testRemoveCollector() public {
         c.addCollector(collector1);
         c.removeCollector(collector1);
-        assertEq( c.collector1(), address(0),"remove an existing collector");
+        assertEq(c.collector1(), address(0), "remove an existing collector");
     }
+
     function testRemoveNotExistCollector() public {
         vm.expectRevert("couldn't remove collector");
         c.removeCollector(collector1);
     }
+
     function testGetBalance() public {
         vm.deal(payable(address(c)), initialValue);
-        uint balance = c.getBalance();
+        uint256 balance = c.getBalance();
         assertEq(balance, initialValue);
     }
 }
